feat: support sql console

+++ backend/pkg/model/request/task.go
@@ -5,7 +5,7 @@ import "mime/multipart"
 type SubmitAppRequest struct {
 	File               *multipart.FileHeader `form:"file" binding:"required"`
 	JdkVersion         int32                 `form:"jdk_version" binding:"required"`
-	AuthenticationType int64                 `form:"authentication_type" binding:"required"`
+	AuthenticationType int32                 `form:"authentication_type" binding:"required"`
 	Username           *string               `form:"username"`
 }
